Update jsdoc comments and rename and reorganize method structure to follow convention
