+++ imports/plugins/included/connectors-shopify/server/index.js
@@ -1,4 +1,4 @@
 import "./collections";
 import "./i18n";
 import "./methods";
-import "./webhooks/endpoints";
+import "./endpoints/webhooks";
