+++ imports/plugins/included/connectors-shopify/server/methods/api/index.js
@@ -1,0 +1,1 @@
+import "./products";
