+++ imports/plugins/included/connectors-shopify/server/methods/import/products.js
@@ -6,14 +6,26 @@ import { Logger } from "/server/api";
 import { check, Match } from "meteor/check";
 
 import { Reaction } from "/server/api";
-import { Packages, Products, Tags, Media } from "/lib/collections";
-import { connectorsRoles } from "../lib/roles";
+import { Products, Tags, Media } from "/lib/collections";
+import { getApiInfo } from "../api/api";
+import { connectorsRoles } from "../../lib/roles";
+
+/**
+ * @file Shopify connector import product method
+ *       contains methods and helpers for setting up and removing synchronization between
+ *       a Shopify store and a Reaction shop
+ * @module connectors/shopify/import/products
+ * location: imports/plugins/included/connectors-shopify/server/methods/import-products.js
+ */
 
 /**
  * Transforms a Shopify product into a Reaction product.
+ * @private
  * @method createReactionProductFromShopifyProduct
  * @param  {object} options { shopifyProduct, shopId, hashtags }
  * @return {[type]} An object that fits the `Product` schema
+ *
+ * @todo consider abstracting private Shopify import helpers into a helpers file
  */
 function createReactionProductFromShopifyProduct(options) {
   const { shopifyProduct, shopId, hashtags } = options;
@@ -60,6 +72,7 @@ function createReactionProductFromShopifyProduct(options) {
 
 /**
  * Transforms a Shopify variant into a Reaction variant.
+ * @private
  * @method createReactionVariantFromShopifyVariant
  * @param  {object} options { shopifyVariant, variant, index, ancestors, shopId }
  * @return {[type]} An object that fits the `ProductVariant` schema
@@ -112,6 +125,7 @@ function createReactionVariantFromShopifyVariant(options) {
 
 /**
  * Finds the images associated with a particular shopify variant
+ * @private
  * @method findVariantImages
  * @param  {Number} shopifyVariantId The variant `id` from shopify
  * @param  {Array} images An array of images from a Shopify product
@@ -124,31 +138,13 @@ function findVariantImages(shopifyVariantId, images) {
 }
 
 /**
- * get Shopify Api Key, Password and Domain from the Shopify Connect package with the supplied shopId or alternatly the active shopId
- * @method getApiInfo
- * @param  {string} [shopId=Reaction.getShopId()] The shopId to get the API info for. Defaults to current shop.
- * @return {object} shopify API connection information
- */
-function getApiInfo(shopId = Reaction.getShopId()) {
-  const { settings } = Packages.findOne({
-    name: "reaction-connectors-shopify",
-    shopId
-  });
-
-  return {
-    apiKey: settings.apiKey,
-    password: settings.password,
-    shopName: settings.shopName
-  };
-}
-
-/**
  * Finds and returns arrays of option values for each of Shopify's option layers
  * returns an object consisting of the following three values:
  * shopifyVariants representing the first option on the shopify product (`option1` in the variant)
  * shopifyOptions representing the second option on the shopify product (`option2` in the variant)
  * shopifyTernary representing the third option on the shopify product (`option3` in the variant)
  * any of these will return undefined if the product does not have an option at that layer.
+ * @private
  * @method getShopifyVariantsAndOptions
  * @param  {object} shopifyProduct The shopify product we are importing
  * @return {object} returns an object consisting of shopifyVariants, shopifyOptions, and shopifyTernaryOptions
@@ -179,6 +175,7 @@ function getShopifyVariantsAndOptions(shopifyProduct) {
 
 /**
  * Transforms a weight in grams to a weight in the shop's default unitsOfMeasure
+ * @private
  * @method normalizeWeight
  * @param  {number} weight weight of the product in grams
  * @return {number} weight of the product in the shop's default unitsOfMeasure
@@ -192,6 +189,7 @@ function normalizeWeight(weight) {
 /**
  * Saves an image from a url to the Collection FS image storage location
  * (default: Mongo GridFS)
+ * @private
  * @method saveImage
  * @param  {string}  url url of the image to save
  * @param  {object}  metadata metadata to save with the image
@@ -208,41 +206,14 @@ function saveImage(url, metadata) {
 
 export const methods = {
   /**
-   * Gets a count of the products from Shopify with the API credentials setup for your store.
-   *
-   * @async
-   * @method shopifyConnect/getProductsCount
-   * @param {object} options An object of options for the shopify API call. Available options here: https://help.shopify.com/api/reference/product#count
-   * @returns {number} Number of products
-   */
-  async "shopifyConnect/getProductsCount"(options) {
-    check(options, Match.Maybe(Object));
-
-    if (!Reaction.hasPermission(connectorsRoles)) {
-      throw new Meteor.Error(403, "Access Denied");
-    }
-
-    const apiCreds = getApiInfo();
-    const shopify = new Shopify(apiCreds);
-    const opts = Object.assign({}, { published_status: "published" }, { ...options });
-
-    try {
-      const count = await shopify.product.count(opts);
-      return count;
-    } catch (err) {
-      Logger.error("something went wrong");
-    }
-  },
-
-  /**
    * Imports products for the active Reaction Shop from Shopify with the API credentials setup for that shop.
    *
    * @async
-   * @method shopifyConnect/getProductsCount
+   * @method connectors/shopify/import/products
    * @param {object} options An object of options for the shopify API call. Available options here: https://help.shopify.com/api/reference/product#index
    * @returns {array} An array of the Reaction product _ids (including variants and options) that were created.
    */
-  async "shopifyConnect/importProducts"(options) {
+  async "connectors/shopify/import/products"(options) {
     check(options, Match.Maybe(Object));
 
     if (!Reaction.hasPermission(connectorsRoles)) {
