+++ imports/plugins/included/connectors-shopify/server/methods/sync/index.js
@@ -1,0 +1,2 @@
+import "./sync";
+import "./orders";
