+++ imports/plugins/included/connectors-shopify/server/methods/webhooks/index.js
@@ -1,0 +1,1 @@
+import "./webhooks";
